requirements.txt
customtkinter>=5.2.0
transformers>=4.35.0
torch>=2.0.0
pillow>=10.0.0
tqdm>=4.66.0
numpy>=1.24.0
python-dotenv>=1.0.0 